committed by bratin

+++ models/activity_budget.js
@@ -4,8 +4,8 @@ module.exports = (sequelize, DataTypes) => {
     activity_id: DataTypes.INTEGER,
     level_type: DataTypes.STRING,
     budget_id: DataTypes.INTEGER,
-    hour: DataTypes.INTEGER,
-    amount: DataTypes.INTEGER,
+    hour: DataTypes.REAL,
+    amount: DataTypes.REAL,
     status: DataTypes.INTEGER,
     activity_goal_id: DataTypes.INTEGER,
     approver_remarks: DataTypes.TEXT
